move content to lib / config dirs
