+++ .eslintignore
@@ -2,3 +2,4 @@ node_modules
 build
 reports
 **/coverage
+aggregated-translations
