--- bin/aggregate-translations.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-
-require('../script/aggregate-translations-cli');
