+++ config/defaultSearchPatterns.js
