+++ lib/aggregate-messages.js
