+++ lib/aggregate-translations.js
@@ -2,12 +2,12 @@ const path = require('path');
 const chalk = require('chalk');
 const fse = require('fs-extra');
 const glob = require('glob');
-const supportedLocales = require('./i18nSupportedLocales');
+const supportedLocales = require('../config/i18nSupportedLocales');
 
 const aggregateMessages = require('./aggregate-messages');
 const writeAggregatedTranslations = require('./write-aggregated-translations');
 const writeI18nLoaders = require('./write-i18n-loaders');
-const defaultSearchPatterns = require('./defaultSearchPatterns');
+const defaultSearchPatterns = require('../config/defaultSearchPatterns');
 
 const isFile = filePath => (fse.existsSync(filePath) && !fse.lstatSync(filePath).isDirectory());
