+++ lib/generate-translation-file.js
