+++ lib/write-aggregated-translations.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const chalk = require('chalk');
-const supportedLocales = require('./i18nSupportedLocales');
+const supportedLocales = require('../config/i18nSupportedLocales');
 const generateTranslationFile = require('./generate-translation-file');
 
 const writeAggregatedTranslations = (aggregatedMessages, locales, fileSystem, outputDir) => {
