+++ lib/write-i18n-loaders.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const chalk = require('chalk');
 const startCase = require('lodash.startcase');
-const supportedLocales = require('./i18nSupportedLocales');
+const supportedLocales = require('../config/i18nSupportedLocales');
 
 const createIntlLoader = (loaderName, locale, format) => {
   if (format === 'es6') {
