+++ package.json
@@ -31,7 +31,7 @@
   "bin": {
     "tt-aggregate-translations": "bin/aggregate-translations.js"
   },
-  "main": "script/aggregate-translations.js",
+  "main": "lib/aggregate-translations-cli.js",
   "dependencies": {
     "chalk": "^2.4.2",
     "commander": "^2.19.0",
